Make the mypy --strict CI gate actually pass (91 errors -> 0)
The previous commit added `mypy --strict env agents` to CI without running it.
It reported 91 errors, so CI would have been red on arrival -- adding a gate
that fails is worse than not adding it, because it trains everyone to ignore a
red build.

Two of the 91 were real staleness, not annotation noise:

- `_apply_action` was still annotated `-> tuple[float, dict]` after the Hard
  Rule 1 refactor changed it to return the physical outcome rather than a
  reward. The annotation had been lying about the return type since that
  commit.
- `_build_trace` was annotated `-> SyntheticSKRQBERTrace`, which the new trace
  mode silently violated by returning a `TraceSKRQBERSource`. Now the
  `SKRQBERTrace` protocol both implement.

The bulk were `Optional` attributes that reset() populates and ~40 sites then
used without narrowing. Rather than sprinkle asserts, the six that have no
meaningful pre-reset value (`_pool_sim`, `_deferral_queue`, `_policy_table`,
`_request_stream`, `_last_pool_state`) are now DECLARED without being bound.
Touching one before reset() raises AttributeError naming the attribute, which
is a better error than `'NoneType' object has no attribute ...`. `_current_mask`
stays Optional, because it is legitimately None before the first reset and is
what `step()` guards on.

Also: `gym.Env` given its type parameters, networkx `Graph[int]` generics,
`yaml.safe_load`'s Any pinned at each config loader, and three genuine
`None`-dereference paths given explicit raises (`_current_floor`,
`_threat_source`/`_threat_scorer`, `REKEY_NOW`'s resolved key type).

Installed types-PyYAML, pandas-stubs and types-networkx.

All four CI gates now pass locally: ruff check, ruff format --check,
mypy --strict on env/ and agents/, and 732 tests.

+++ agents/mpc_oracle.py
@@ -184,7 +184,7 @@ class MPCOracle:
         policy sees this too, via `state["pool_fill"]`."""
         if self.env is None:
             return 0.0
-        return self.env._pool_sim.fill / self.env._bits_per_hybrid_draw
+        return float(self.env._pool_sim.fill / self.env._bits_per_hybrid_draw)
 
     def act(self, state: StateDict, mask: ActionMask) -> Action:
         demand, refill = self.peek_future()
